Refactored Code
Did some refactoring to make it so that controller can be substituted easily. Still needs some minor refactoring tweaks, and there might need to be a few small modifications when the controller is added. Also still need accurate positioning system implementation.
