+++ message_creator.py
@@ -3,27 +3,20 @@
 import rospy
 
 # simple if else for low battery warning
-def create_message(twist, batteryLevel, posX, posY, orient_x, orient_y, orient_z, orient_w):
+def create_message(move_command, batteryLevel, distance, angle):
+    
     if(batteryLevel>=10):
         msg = """
         Sending commands to turtlebot...
         --------------------------------
-        Linear velocities
-        X: %f
-        Y: %f
+        Linear velocity: %f
         Angular Velocity (CCW): %f
         --------------------------------
         Odometry
 
-        Position:
-        X: %f
-        Y: %f
+        Distance travelled: %f
 
-        Orientation:
-        x: %f
-        y: %f
-        z: %f
-        w: %f
+        Angle (degrees): %f 
         --------------------------------
         Estimated battery level: %f/100
         --------------------------------
@@ -35,35 +28,23 @@ def create_message(twist, batteryLevel, posX, posY, orient_x, orient_y, orient_z
         msg = """
         Sending commands to turtlebot...
         --------------------------------
-        Linear velocities
-        X: %f
-        Y: %f
-        Angular Velocity: %f
+        Linear velocity: %f
+        Angular Velocity (CCW): %f
         --------------------------------
         Odometry
-        X Position: %f
-        Y Position: %f
-        --------------------------------
-        Orientation (IMU)
-        x: %f
-        y: %f
-        z: %f
-        w: %f
+
+        Distance travelled: %f
+
+        Angle (degrees): %f
         --------------------------------
         Estimated battery level: %f/100
-        BATTERY LEVEL LOW. CHARGE SOON
+        BATTERY LEVEL LOW. CHARGE SOON.
         --------------------------------
         
-
-
-
-
-
         --------------------------------
 
         """
 
     # write the above message to the console with velocity, position, battery level values
-    rospy.loginfo(msg, round(twist.linear.x, 2), round(twist.linear.y, 2), round(twist.angular.z, 2), round(posX, 2), round(posY, 2),
-    round(orient_x, 2), round(orient_y, 2), round(orient_z, 2), round(orient_w, 2), batteryLevel)
+    rospy.loginfo(msg, round(move_command.linear.x, 2), round(move_command.angular.z, 2), distance.x, angle, batteryLevel)
