+++ pid_control.py
@@ -1,14 +1,13 @@
 # class for implementing PID control on the turtlebot
 # evaluates a PID controller given current state
 
-import math
 import time
 
 
 class pid_controller(object):
 
-    def __init__(self, p_gain, d_gain, i_gain, i_min, i_max):
-        self._set_gains(p_gain, d_gain, i_gain, i_min, i_max)
+    def __init__(self, p_gain, d_gain, i_gain):
+        self._set_gains(p_gain, d_gain, i_gain)
         self._reset()
         self.current_time = time.time()
 
@@ -21,20 +20,12 @@ class pid_controller(object):
         self._last_time = None
     
     
-    def _set_gains(self, p_gain, d_gain, i_gain, i_min, i_max):
+    def _set_gains(self, p_gain, d_gain, i_gain):
         # setter for proportional, derivative, and integral components
         # of a the controller with a built-in check for integral gain bounds
         self._p_gain = p_gain
         self._d_gain = d_gain
-
-        if i_gain > i_max:
-            i_gain = i_max
-        elif i_gain < i_min:
-            i_gain = i_min
-
         self._i_gain = i_gain
-        self._i_min = i_min
-        self._i_max = i_max
 
     
     def _get_gains(self):
